removed unusded files
