+++ packages/domain/src/aggregate.ts
@@ -43,8 +43,8 @@
 
 import {ILogger} from "@mojaloop/logging-bc-public-types-lib";
 import { IMessageProducer } from "@mojaloop/platform-shared-lib-messaging-types-lib";
-import { GetParticipantError, GetPartyError, NoSuchParticipantError, NoSuchPartyError, UnableToAssociateParticipantError, UnableToAssociatePartyError, UnableToDisassociateParticipantError, UnableToDisassociatePartyError, UnableToGetOracleError, UnableToGetOracleProviderError } from "./errors";
-import { ILocalCache, IOracleFinder, IOracleProvider} from "./interfaces/infrastructure";
+import { GetParticipantError, GetPartyError, NoSuchParticipantError, NoSuchParticipantFspIdError, NoSuchPartyError, UnableToAssociateParticipantError, UnableToAssociatePartyError, UnableToDisassociateParticipantError, UnableToDisassociatePartyError, UnableToGetOracleError, UnableToGetOracleProviderError } from "./errors";
+import { ILocalCache, IOracleFinder, IOracleProvider, IParticipantService} from "./interfaces/infrastructure";
 import { IParticipant, IParty, LocalCacheKeyPrefix } from "./types";
 
 export class AccountLookupAggregate {
